#128 Service interfaces added
