+++ src/main/java/juja/microservices/gamification/service/KeeperService.java
@@ -1,21 +1,12 @@
 package juja.microservices.gamification.service;
 
-import juja.microservices.gamification.dao.KeeperRepository;
 import juja.microservices.gamification.entity.KeeperDTO;
-import org.springframework.stereotype.Service;
 
-import javax.inject.Inject;
 import java.util.List;
 
-/**
- * @author Vadim Dyachenko
+/*
+ * @author Benjamin Novikov, November 2017, task #
  */
-@Service
-public class KeeperService {
-    @Inject
-    private KeeperRepository keeperRepository;
-
-    public List<KeeperDTO> getKeepers() {
-        return keeperRepository.getKeepers();
-    }
+public interface KeeperService {
+    List<KeeperDTO> getKeepers();
 }
